+++ src/main/java/dandi/dandi/pushnotification/application/service/WeatherPushNotificationScheduler.java
@@ -7,6 +7,7 @@ import dandi.dandi.pushnotification.application.port.out.persistence.PushNotific
 import dandi.dandi.pushnotification.application.port.out.webpush.WebPushManager;
 import dandi.dandi.pushnotification.application.service.message.WeatherPushNotificationMessageGenerator;
 import dandi.dandi.pushnotification.domain.PushNotification;
+import dandi.dandi.pushnotification.domain.PushNotificationSource;
 import dandi.dandi.pushnotification.domain.RetryableWeatherPushNotification;
 import dandi.dandi.weather.application.port.out.WeatherForecastInfoManager;
 import dandi.dandi.weather.application.port.out.WeatherForecastResponse;
